Add APPSIGNAL_API_KEY env var to set api key
Previously, the endpoint to post to was set through a variable named
ENDPOINT, which contained the whole URL, including the path and API
key.

This patch splits up the configuration, allowing the user to set an
APPSIGNAL_API_KEY without having to know anything about the
internals. The endpoint defaults to appsignal-endpoint.net, but can be
overridden using APPSIGNAL_ENDPOINT.