Ignore nameless entities
Do not emit metrics for a node, pod or volume if it doesn't have a
name. In practice, this only happens for volumes.

When a metric is missing, instead of emitting a metric with a value
of zero, do not emit a metric at all.

Re-structure the metric extraction functions to be able to test the
way in which the overall JSON is traversed, rather than just how
metrics are extracted from each section of it.