Crash if `APPSIGNAL_API_KEY` is not set
A panic in an asychronous task becomes a `JoinError` when awaited.
As we now print the errors returned by `run()`, they will no longer
cause the thread to panic. This behaviour is correct for errors
related to the network requests, which can be assumed to be transient,
but the process should actually crash if `APPSIGNAL_API_KEY` is not
set.

Move the creation of the URL that the metrics are sent to to a
separate function, as to call it in `main()` instead of `run()`,
causing the process to crash if the URL fails to be created because
`APPSIGNAL_API_KEY` is not set.