Improve the release process
I noticed we ship the image with the `rust:alpine` image as a base. This
means we ship the entire Rust language with our releases as well. Our
image size is therefore 559.38 MB according to Docker Hub.
This was because the `Dockerfile` did two things at once:

- Build the project with Cargo
- And be the thing we publish to Docker Hub

Rewrite the build and publish process to only ship a small image with an
alpine base that contains the `appsignal-kubernetes` executable without
the Rust language.

The `Dockerfile` will now only have the alpine base, and uses the Docker
buildx TARGETARCH ARG to determine which executable release it needs to
copy onto the image.

Rewrite the Makefile to a Rakefile as it allows for more structure in
the more complex release process. I copied a bunch of the agent release
Rakefiles, but only the minimum. It's a bit more verbose but easier to
follow.

I didn't optimize this process too much, so it might be slow to build,
but we can tweak the Docker container volume settings if needed.