Reduce published image size
I noticed we ship the image with the `rust:alpine` image as a base. This
means we ship the entire Rust language, all dependencies, and
compilation caches and artifacts with our releases as well. Our image
size is therefore 559.38 MB according to Docker Hub. This was because
the `Dockerfile` did two things at once:

- Build the project with Cargo
- And be the thing we publish to Docker Hub

Use a multi-stage build to not include the entire Rust language base
image in the published image, reducing the image size from 559.38 MB to
around 7 MB (including the size reducing improvements from PR #6).

The published image will only have the alpine image as a base. It
includes the build artifact from the previous stage by way of copying it
from the first stage to the second stage. This works for the multiple
platforms we publish.