Tag volume metrics by pod
Volume names are only guaranteed to be unique per-pod. While a volume
can be shared between pods, the "volume name" used to declare its
metrics is not the name of the `PersistentVolume` object, but the
name with which the `PersistentVolumeClaim` is defined in the pod
definition -- as such, in that scenario, we report the same metrics
about the volume several times, once for each pod that has a claim
to the volume, with a different pod tag.

Removed a test asserting that repeated metrics are removed, as there
is no feasible scenario in which that happens.