Use independent cache mounts per platform
When building for several platforms, both builds attempt to use the
same `target` cache volume. Not only does this essentially force the
builds to be sequential, as `cargo` takes a lock on the folder, but
the artifacts compiled for one platform cannot be re-used for the
other in future builds, in practice forcing the build to start from
scratch as if no cache were present.

Use the `id` mount attribute and the `$TARGETPLATFORM` environment
variable to have different cache volumes mounted in the same path
for each platform.