Create a new release with Mono
Call mono from the cloned location to create a new version.

It will create a version number based on the changesets and write it to
the `Cargo.toml` using the `script/write_version` script.

I've supplied it with the `--no-git` and `--no-package-push` flags
to make sure it will not make a Git commit and tag for this release. We
will do that manually later.

The `--yes` flag will auto accept any interactive prompts so that it
will not hang in this workflow.