Commit and tag the new release
After Mono has created a new version number and updated the changelog,
commit the changes and tag the commit for the new version.

Fetch the generated version number using GitHub Actions 'outputs' rather
than running `script/read_version` all the time so it's easy to
reference in later steps in the workflow.