Add Rake task to install cross
I'm moving the build steps to a Rakefile, because the code to cross
compile and build the release image gets more complex than I'm
comfortable in maintaining a Makefile.

The first step is to install cross and ensure it's always using
the same version.

I've copied the command runner from our agent project and updated it
a bit.