Add project build tasks
These tasks will build the release artifacts that will be copied
to the release image. They have to be built separately because
the emulation on GitHub Actions runners is so slow it takes 55
minutes to build the ARM target.

This follows the setup we have in our agent project. It first builds the
release artifacts separately and then copies them onto the Docker image.
That last step will be done in a future commit.

https://appsignal.slack.com/archives/C06R4337AJ1/p1727689745049589