Build the release image locally
Rewrite the make command to a Rake task to test the release image
locally.

Instead of making the Docker image with a two stage build step, copy the
release artifact onto the image. This significantly reduces the build
time, as building this image for the ARM platform takes 55 minutes on
GitHub Actions.

I've given the locally built image the tag 'local' so that it doesn't
overwite any pulled release images to avoid confusion about what image
tag someeone is running.