Avoid cloning metrics into keys
In #13, the `.to_key` was modified to take `&self` as an argument
and clone it, following Clippy's (correct) recommendation for
`.to_key` methods to take `&self` as an argument.

However, we have no need in our codebase to use the metric after
converting it to a key, so this incurs a needless clone, which
could always be done explicitly if needed. Undo the change to the
method's signature, but rename it to `.into_key` to stay in line
with Clippy's `wrong_self_convention` lint rule.