Replace metrics HashSet with a namespaced HashMap
In preparation of namespaced data, replace the existing HashSet with a
HashMap, which will hold a key per namespace. For now, place all data
under the "default" namespace. Then, loop over each namespace in the
out variable when sending metrics to AppSignal, and send all matrics
for one namespace as one request.