Remove quotes from metric node names
Previously, the returned nodeName was converted to a string with
to_string(), which turns the JSON representation into a string. This
includes quotes to indicate the value is indeed a string.

This patch uses as_str(), which returns an option. The node name is
only set if a str is actually returned.