Info print configuration when starting
When the integration starts, info log a startup message with the used
configuration for debug purposes.

This patch introduces the Config struct, which has a from_env()
function that takes the configuration values from the environment
variables. This function is now used on startup and when building the
request URL when sending metrics.