Build each target on CI
To make sure any changes don't break the release builds, let's build
them in the CI.

That way, we don't find out the build is broken when we try to run the
release workflow.

I've reverted the `cargo test` commands to just run `cargo test` on the
Ubuntu runner for now as we do in our other projects too.

We may want to revisit this, but let's first add the release build test
steps.