Fix Git remote in publish workflow
The Helm chart releaser needs the Git remote to be a HTTPS URL.

We check out the repository with SSH using a deploy key at the start of
the workflow, because any other workflows that depend on `push` events
to trigger them need to be pushed with a deploy key.

Update the remote so it works with the Helm releaser again.

[skip changeset]
[skip ci]