Add pod restart count and uptime
Adds the pod restart count and the uptime in seconds to the metrics
sent to AppSignal for each pod.

Note that, although the `chrono` dependency is now added as a direct
dependency of the crate, it was already an indirect dependency of it,
and the dependency version remains the same.