Report top-level owners for pods
For each pod, report the top-level owner resources for those pods,
that is, the transitive owners which are not owned by any other
resources. Note that a pod may have more than one owner, and therefore
it can also have more than one top-level owner.

Update the `kube` crate to version 1.1.0 and the `k8s-openapi` crate
to version `0.25.0`. Use the "earliest" feature instead of "latest"
to support most API server versions, though there is no significant
difference in support regardless.

Update the Rust version to 1.82.0 to be able to use a recent version
of the `kube` crate.

Implement a cache for ownership resolution and API discovery. The
discovery results are kept between runs, but can be refreshed if
resources from new APIs are referenced. The ownership cache is reset
during each run, as the ownership relationships can change at any
time.

Improve trace logging. Do not `panic!` when AppSignal or Kubernetes
are unreachable.

Update the deployment YAML to provide read-only permissions to all
API groups and resources -- this is necessary due to the owner
traversal logic needing to be able to retrieve arbitrary resources.