Implement an actual Helm chart
Our existing Helm chart is just a copy and paste of the existing
`deployment.yaml` for `kubectl`, with some unrelated broken files
alongside it.

Replace this with templates that actually interpolate the values in
the chart. Rather than having the release process copy-paste the
`deployment.yaml` from the root into the chart, do the opposite: use
Helm's template as the source of truth to generate `deployment.yaml`
from.