Report metrics when pods have no stats
When a pod metric has no stats (CPU, memory... as reported by the node
stats summary endpoint) report it anyway. Do this by storing the
pods stats while traversing the stats JSON, but emitting metrics for
all pods, as reported by the Kubernetes API, not just for those
listed in the stats endpoint.