Fix namespace creation in Helm
The Helm chart included a `namespace.yaml` definition, which was
useful so that the `deployment.yaml` generated from the Helm chart
would contain a namespace. However, this breaks Helm's `--namespace`
and `--create-namespace` flags, which are the preferred way to
install a Helm chart on a specific namespace, and are what we
recommend on our README.

Turn the existing `managed` flag into a `kubectl` flag that determines
whether we're compiling the Helm template to be used via `kubectl`.
This flag is off in the default values, but when it's on, it will
cause the namespace to be present and remove Helm-related labels.